+++ chat/chat-window/src/ChatWindow/ChatWindow.types.ts
@@ -5,4 +5,4 @@ import { DarkModeProps, HTMLElementProps } from '@leafygreen-ui/lib';
 export interface ChatWindowProps
   extends Omit<HTMLElementProps<'div'>, 'title'>,
     DarkModeProps,
-    TitleBarProps {}
+    Partial<TitleBarProps> {}